pyproject.toml
[project]
name = "rtflite"
version = "1.0.2"
description = "Lightweight RTF composer for Python"
authors = [
    { name = "Yilong Zhang", email = "elong0527@gmail.com" },
    { name = "Nan Xiao", email = "me@nanx.me" },
]
dependencies = [
    "pydantic>=2.0.0",
    "pillow>=8.0.0",
    "narwhals>=1.0.0",
    "polars>=1.0.0",
]
readme = "README.md"

classifiers = [
    "Development Status :: 3 - Alpha",

    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Healthcare Industry",
    "Intended Audience :: Information Technology",

    "Topic :: Text Processing",
    "Topic :: Office/Business :: Office Suites",

    "License :: OSI Approved :: MIT License",

    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]

requires-python = ">= 3.10"

[project.urls]
Homepage = "https://pharmaverse.github.io/rtflite/"
Documentation = "https://pharmaverse.github.io/rtflite/"
Repository = "https://github.com/pharmaverse/rtflite"
Issues = "https://github.com/pharmaverse/rtflite/issues"
Changelog = "https://github.com/pharmaverse/rtflite/blob/main/CHANGELOG.md"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.3.3",
    "pytest-cov>=6.0.0",
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.5.45",
    "mkdocstrings-python>=1.12.2",
    "markdown-exec[ansi]>=1.11.0",
    "ruff>=0.7.4",
    "isort>=5.13.2",
    "mypy>=1.17.1",
    "matplotlib>=3.10.5",
    "pandas>=2.2.3",
    "pyarrow >= 21.0.0; python_version < '3.14'",
    "pyarrow >= 22.0.0; python_version >= '3.14'",
]

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]
